+++ src/main/java/kusitms/jangkku/domain/user/application/UserServiceImpl.java
@@ -1,6 +1,5 @@
 package kusitms.jangkku.domain.user.application;
 
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
 import kusitms.jangkku.domain.interest.dao.InterestRepository;
 import kusitms.jangkku.domain.interest.domain.Interest;
@@ -11,10 +10,10 @@ import kusitms.jangkku.domain.token.domain.RefreshToken;
 import kusitms.jangkku.domain.user.dao.UserRepository;
 import kusitms.jangkku.domain.user.domain.User;
 import kusitms.jangkku.domain.user.dto.UserDto;
-import kusitms.jangkku.domain.userinterest.dao.UserInterestRepository;
-import kusitms.jangkku.domain.userinterest.domain.UserInterest;
-import kusitms.jangkku.domain.userkeyword.dao.UserKeywordRepository;
-import kusitms.jangkku.domain.userkeyword.domain.UserKeyword;
+import kusitms.jangkku.domain.user.dao.UserInterestRepository;
+import kusitms.jangkku.domain.user.domain.UserInterest;
+import kusitms.jangkku.domain.user.dao.UserKeywordRepository;
+import kusitms.jangkku.domain.user.domain.UserKeyword;
 import kusitms.jangkku.global.util.CookieUtil;
 import kusitms.jangkku.global.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
