+++ src/main/java/kusitms/jangkku/domain/user/dao/UserInterestRepository.java
@@ -1,6 +1,6 @@
-package kusitms.jangkku.domain.userinterest.dao;
+package kusitms.jangkku.domain.user.dao;
 
-import kusitms.jangkku.domain.userinterest.domain.UserInterest;
+import kusitms.jangkku.domain.user.domain.UserInterest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
