+++ src/main/java/kusitms/jangkku/domain/user/dao/UserKeywordRepository.java
@@ -1,6 +1,6 @@
-package kusitms.jangkku.domain.userkeyword.dao;
+package kusitms.jangkku.domain.user.dao;
 
-import kusitms.jangkku.domain.userkeyword.domain.UserKeyword;
+import kusitms.jangkku.domain.user.domain.UserKeyword;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
