+++ src/main/java/kusitms/jangkku/domain/user/domain/UserInterest.java
@@ -1,8 +1,7 @@
-package kusitms.jangkku.domain.userkeyword.domain;
+package kusitms.jangkku.domain.user.domain;
 
 import jakarta.persistence.*;
 import kusitms.jangkku.domain.interest.domain.Interest;
-import kusitms.jangkku.domain.keyword.domain.Keyword;
 import kusitms.jangkku.domain.user.domain.User;
 import lombok.AccessLevel;
 import lombok.Builder;
@@ -12,24 +11,24 @@ import lombok.NoArgsConstructor;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Table(name = "users_keywords")
-public class UserKeyword {
+@Table(name = "users_interests")
+public class UserInterest {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "users_keywords_id")
+    @Column(name = "users_interests_id")
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "users_id", foreignKey = @ForeignKey(name = "users_keywords_fk_users_id"))
+    @JoinColumn(name = "users_id", foreignKey = @ForeignKey(name = "users_interests_fk_users_id"))
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "keywords_id", foreignKey = @ForeignKey(name = "users_keywords_fk_keywords_id"))
-    private Keyword keyword;
+    @JoinColumn(name = "interests_id", foreignKey = @ForeignKey(name = "users_interests_fk_interests_id"))
+    private Interest interest;
 
     @Builder
-    public UserKeyword(User user, Keyword keyword) {
+    public UserInterest(User user, Interest interest) {
         this.user = user;
-        this.keyword = keyword;
+        this.interest = interest;
     }
 }
